% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cv_AUC.R
\name{ci_cv_AUC}
\alias{ci_cv_AUC}
\title{Calculates cross validated AUC with 95\% confidence intervals for a given caret model object}
\usage{
ci_cv_AUC(model)
}
\arguments{
\item{model}{model object obtained with \code{\link[caret:train]{caret::train()}}.}
}
\description{
Calculates cross validated AUC with 95\% confidence intervals for a given caret model object
}
