% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coefs.R
\name{extract_coefs}
\alias{extract_coefs}
\title{Extract coefficients from model}
\usage{
extract_coefs(modelList)
}
\arguments{
\item{modelList}{containing the results from the modelling process obtained with \code{\link[=ml_eval]{ml_eval()}}}
}
\description{
Extracts the name of coefficients that are present in each iteration of k-fold cross validation to get
statistics of feature importance. Parameters which are included in many of the different splits are considered
important features.
}
