% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_cvAUC.R
\name{ls_cvAUC}
\alias{ls_cvAUC}
\title{Prepare data for input into \code{\link[cvAUC:cvAUC]{cvAUC::cvAUC()}}}
\usage{
ls_cvAUC(modelList)
}
\arguments{
\item{modelList}{containing the results from the modelling process obtained with \code{\link[=ml_eval]{ml_eval()}}}
}
\description{
Function to split data in a way that the function cvAUC can use it for calculation of ROC, se and confidence intervals
}
