% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_eval.R
\name{ml_eval}
\alias{ml_eval}
\title{Conduct nested cv using the calc_model_metrics function in this package}
\usage{
ml_eval(
  modelMatrix,
  y,
  k.outer = 10,
  rep.outer = 10,
  k.inner = 10,
  rep.inner = 5,
  train.method = "glmnet",
  cv.method = "repeatedcv",
  metric = "ROC",
  tuneGrid = expand.grid(alpha = 1, lambda = seq(0.01, 0.2, by = 0.01)),
  ...
)
}
\arguments{
\item{modelMatrix}{a model matrix containing the predictor variables. Preferably obtained through the \code{\link[stats:model.matrix]{stats::model.matrix()}} function.}

\item{y}{vector containing outcome variable.}

\item{k.outer}{define number of folds for k-fold cross validation in the outer loop}

\item{rep.outer}{for repeated k-fold cross-validation only: the number of complete sets of folds to compute in the outer loop}

\item{k.inner}{number of folds for k-fold cross validation in the inner loop}

\item{rep.inner}{number of repeats for repeated k-fold cross validation in the inner loop}

\item{train.method}{A string specifying which classification or regression model to use. Possible values are found using names(caret::getModelInfo()).}

\item{cv.method}{The resampling method: "boot", "boot632", "optimism_boot", "boot_all", "cv", "repeatedcv", "LOOCV", "LGOCV" (for repeated training/test splits), "none" (only fits one model to the entire training set), "oob" (only for random forest, bagged trees, bagged earth, bagged flexible discriminant analysis, or conditional tree forest models), timeslice, "adaptive_cv", "adaptive_boot" or "adaptive_LGOCV".}

\item{metric}{A string that specifies what summary metric will
be used to select the optimal model. By default, possible values
are "RMSE" and "Rsquared" for regression and "Accuracy" and
"Kappa" for classification. If custom performance metrics are
used (via the \code{summaryFunction} argument in
\code{\link[caret]{trainControl}}, the value of \code{metric} should
match one of the arguments. If it does not, a warning is issued
and the first metric given by the \code{summaryFunction} is
used. (NOTE: If given, this argument must be named.)}

\item{tuneGrid}{A data frame with possible tuning values. The
columns are named the same as the tuning parameters. Use
\code{\link[caret]{getModelInfo}} to get a list of tuning parameters
for each model or see
\url{http://topepo.github.io/caret/available-models.html}.
(NOTE: If given, this argument must be named.)}

\item{...}{additional arguments passed on to function \code{\link[=calc_model_metrics]{calc_model_metrics()}}}
}
\description{
This function takes a model Matrix as input and splits it into a defined number of folds with a specified number of repetitions.
Afterwards a nested cv is conducted with the \code{calc_model_metrics} function contained in this package. The
results are given as a list of the training process, model coefficients and model evaluation. To accelerate the modelling
process parallel computing is activated.
}
\seealso{
\code{\link[=calc_model_metrics]{calc_model_metrics()}} for more arguments.
}
