% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_model_ci.R
\name{rbind_model_ci}
\alias{rbind_model_ci}
\title{Calculate inner cvAUC and outer cvAUC and display in dataframe}
\usage{
rbind_model_ci(modelList, train.cv = "train.cv", rep.outer)
}
\arguments{
\item{modelList}{list containing the results from the modelling process obtained with \code{\link[=ml_eval]{ml_eval()}}.}

\item{train.cv}{name of the object containing cvAUC results of the training process. Character string.}

\item{rep.outer}{number of repeats in the outer loop}
}
\description{
This function calculates the cvAUC of the inner loop and outer loop as well as the respective
95\% confidence intervals.
}
