% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_model.R
\name{unlist_model}
\alias{unlist_model}
\title{convert metrics from training data list to vector}
\usage{
unlist_model(modelList, metric, element, rep.outer)
}
\arguments{
\item{modelList}{model results generated with \code{\link[=ml_eval]{ml_eval()}}. List object.}

\item{metric}{parameter to extract from the modelList.}

\item{element}{list element from which the metric parameter will be extracted.}

\item{rep.outer}{number of repeats in the outer loop}
}
\description{
This function takes the results generated with the \code{\link[=ml_eval]{ml_eval()}} function and converts the results list
into a vector depending on the metric and list element.
}
